refactor : 컴포넌트 패키지 정리
