+++ feature/home/src/main/java/com/wepli/home/component/banner/BannerComponent.kt
@@ -1,4 +1,4 @@
-package com.wepli.home.screen.relaylist
+package com.wepli.home.component.banner
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -8,8 +8,6 @@ import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.wepli.home.component.WePLiBanner
-import com.wepli.home.component.WePLiBannerType
 
 @Composable
 fun WePLiBannerLayout() {
