+++ feature/home/src/main/java/com/wepli/home/screen/HomeScreen.kt
@@ -3,24 +3,14 @@ package com.wepli.home.screen
 import android.annotation.SuppressLint
 import android.widget.Toast
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
@@ -30,32 +20,20 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import appbar.HomeAppBar
 import appbar.WepliAppBar
 import com.wepli.feature.home.R
-import com.wepli.home.component.PlayListCoverItem
-import com.wepli.home.component.WePLiBanner
-import com.wepli.home.component.WePLiBannerType
 import com.wepli.home.mvi.HomeEffect
 import com.wepli.home.mvi.HomeIntent
 import com.wepli.home.mvi.HomeUiState
-import com.wepli.home.screen.relaylist.ArtistLayout
-import com.wepli.home.screen.relaylist.RelaylistPagerLayout
-import com.wepli.home.screen.relaylist.WePLiBannerLayout
-import com.wepli.home.screen.relaylist.WePLiChartLayout
-import com.wepli.home.screen.relaylist.WePLiPlaylistLayout
+import com.wepli.home.component.artist.ArtistLayout
+import com.wepli.home.component.relaylist.RelaylistPagerLayout
+import com.wepli.home.component.banner.WePLiBannerLayout
+import com.wepli.home.component.chart.WePLiChartLayout
+import com.wepli.home.component.playlist.WePLiPlaylistLayout
 import com.wepli.home.viewmodel.HomeViewModel
 import com.wepli.shared.feature.mock.artistMockData
 import com.wepli.shared.feature.mock.musicMockData
 import com.wepli.shared.feature.mock.recommendPlaylistMockData
 import com.wepli.shared.feature.mock.relaylistUiMockData
-import com.wepli.shared.feature.uimodel.artist.ArtistUiData
-import com.wepli.uimodel.music.ChartMusicUiData
-import compose.MeasuredHeightContainer
-import custom.ArtistProfileListItem
-import custom.MusicItem
-import custom.MusicItemType
-import custom.OneLineTitle
-import custom.TwoLineTitle
 import dev.chrisbanes.haze.hazeSource
-import model.playlist.RecommendPlaylist
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
 import theme.LocalHazeState
